Includes self on GetImplementedInterface calls to interface
For situations where an interface is passed to GetImplementedInterfaces
this adds the actual interface to the list of returned values.

This allows for situations such as registering the return value of a
factory method. Prior to this commit to resolve that type you'd have to
use AsSelf() rather than AsImplementedInterfaces().

Resolves issue #770