Fix crash on IOS with linker
On IOS, when the mono linker is enabled (and it’s enabled by default
and recommended by Xamarin) it strips out all parts of the framework
which are not in use. The linker does not support reflection, however,
therefore it does not recognize the call to
'GetRuntimeMethod(nameof(Convert.ChangeType)' (which is called in
ConstructorParameterBinding -> ConvertPrimitiveType) as using the
'Convert.ChangeType' method and strips it away unless it's explicitly
used somewhere else, crashing Autofac if that method is called (null
reference exception as ‘GetRuntimeMethod’ returns null).
Added an explicit call to Convert.ChangeType to let the linker know
that we need this method and with that making sure Autofac works on IOS.