LifetimeScope: Properly dispose of allocated registries.
When a child scope is created, along with its registry, we need to keep track
of what needs to be disposed of later. The registry is part of those things.

The registry passed in the constructor doesn't always fall in that category
(think root scope), and hence the registry is added in the BeginLifetimeScope
methods.

Issue #746