Breaking: DRE is no longer serializable
https://github.com/dotnet/docs/issues/34893
.NET has long been deprecating BinaryFormatter and Serializable items.
Analyzers as of .NET 8 are warning about obsolete serialization
constructors. Given most of the underpinnings of this mechanism are also
being deprecated, it seemed reasonable to include deprecation for that in
Autofac.

For projects that continue to use AppDomains and need to serialize
exceptions over the wire, it's recommended to not upgrade Autofac to
this version.